additional complexity and logistic step calculations
added a function that considers that you wont need to craft a component at different station but you rather would get it provided by the same station. so when you need like gears for 3 different other components the function would only list the logistic steps to craft gears once but adding one logistic step for each time the gear is required as material.